test/backend/requirements.txt
fastapi
uvicorn
pymongo
PyJWT